requirements.txt
hydralit_components==1.0.10
st_annotated_text==4.0.0
streamlit==1.22.0
